Edits

+++ RDFTMinigame/EngineLib/EngineCommon.cpp
@@ -1,6 +1,6 @@
 #include "EngineCommon.h"
 
-GLVECTOR2 VectorOf(float x, float y) {
+GLVECTOR2 VectorOf(double x, double y) {
 	GLVECTOR2 pos;
 	pos.x = x;
 	pos.y = y;
